GeneratorContextForMySql test
