+++ src/Dapper-Layers-Generator/Dapper-Layers-Generator.Core/Generators/GeneratorContextForProvider.cs
@@ -54,7 +54,7 @@ namespace Dapper_Layers_Generator.Core.Generators
                 // Specific for DB provider {{DbProviderString}}
                 // Author: {{_settings.AuthorName}}
                 // Context name: {{_settings.DbContextClassName}}
-                // Generated: {{_settings.GenerationTimestamp}}
+                // Generated: {{_settings.GenerationTimestamp.ToString("yyyy-MM-dd HH:mm:ss")}} UTC
                 // WARNING: Never change this file manually (re-generate it)
                 // =================================================================
