+++ test/Dapper-Layers-Generator.Test/Generators/GeneratorContextForBase_Tests.cs
@@ -4,26 +4,15 @@ using Moq;
 using System.Text.Json;
 using Dapper_Layers_Generator.Data.POCO;
 using Dapper_Layers_Generator.Core.Converters;
+using Dapper_Layers_Generator.Test.Generators;
 
 namespace Dapper_Layers_Generator.Core.Generators.Tests
 {
-    public class GeneratorContextForBase_Tests
+    public class GeneratorContextForBase_Tests : GeneratorBaseTest
     {
-
-        private Mock<IReaderDBDefinitionService> _mockDbDefinitions;
-        private SettingsGlobal _settings;
-
-        public GeneratorContextForBase_Tests()
+        public GeneratorContextForBase_Tests() : base()
         {
-            //Fake DB
-            var jsonDB = ResourceTool.Read("Dapper_Layers_Generator.Test/Data/schema.json");
-            var schemas = JsonSerializer.Deserialize<List<Schema>>(jsonDB) ?? throw new NullReferenceException("Cannot test without JSON DB");
-            _mockDbDefinitions = new Mock<IReaderDBDefinitionService>();
-            _mockDbDefinitions.Setup(x => x.SchemaDefinitions).Returns(schemas);
 
-            //Standard settings
-            var settings = ResourceTool.Read("Dapper_Layers_Generator.Test/Data/baseconfig.json");
-            _settings = JsonSerializer.Deserialize<SettingsGlobal>(settings) ?? throw new NullReferenceException("Cannot test without standard settings");
         }
 
         [Fact()]
